ltl2ba warnings fix

+++ ltl2ba/alternating.c
@@ -433,7 +433,7 @@ void mk_alternating(Node *p) /* generates an alternating automaton for p */
     getrusage(RUSAGE_SELF, &tr_fin);
     timeval_subtract (&t_diff, &tr_fin.ru_utime, &tr_debut.ru_utime);
     fprintf(tl_out, "\nBuilding and simplification of the alternating automaton: %lli.%06ld",
-            (long long int)t_diff.tv_sec, t_diff.tv_usec);
+            (long long int)t_diff.tv_sec, (long int)t_diff.tv_usec);
     fprintf(tl_out, "\n%i states, %i transitions\n", astate_count, atrans_count);
   }
